source_code/scorehive_server/requirements.txt
asgiref==3.6.0
coverage==7.2.6
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
